refactor(alerts): split panel subcomponents into own files per review
CodeRabbit guideline findings on #649: PanelTabsList moves to its own
file, and expandSignal.ts splits into SectionsExpandContext.ts,
useSectionsExpandBroadcast.ts and CollapsibleSection.types.ts so file
names match their exports.
