+++ apps/client/src/components/Alerts/AlertDetails/CollapsibleSection/SectionsExpandContext.ts
@@ -1,0 +1,4 @@
+import { createContext } from 'react';
+import { SectionsExpandSignal } from './CollapsibleSection.types';
+
+export const SectionsExpandContext = createContext<SectionsExpandSignal | null>(null);
